store: deleted config.go. removed Config.DefAge. moved moveToTrash code to top..
